Delete monkeypatch_guard.rb
It's no longer being used.

It's a good idea, but it's incomplete. We should resurrect this idea,
but instead of not multi-running when a single file is changed within a
short time span (e.g. while the spec is running) also don't run multiple
times just because multiple different files were changed.