Move global gems from install.sh to Gemfile
I think that this will be a little bit better/easier of a way to manage
the set of gems that I want to have globally available. The fun thing
about Ruby is that installing a gem at all (even via a specific repo's
Gemfile) causes the gem's executables to be globally available. In
general, that might or might not be a good thing, but, in this case,
that happens to be just what we want.

I guess that another notable advantage of this approach is that I can
set up Dependabot to update the gems in this Gemfile, so I will be
notified when new versions of these gems are released, and then I can
pull down the updated gems by merging the dependabot PRs, updating my
local git branch, and bundle installing.