[zshrc] Source fzf differently on Linux vs MacOS
In 64ec725efd0969384492148706119ca99ef3ec3e, we switched from `source
~/.fzf.zsh` to `source <(fzf --zsh)`, which is the way to source the
latest versions of fzf. However, on my Mac, I cannot update to the
latest version of fzf, and so that approach raises an error (`unknown
option: --zsh`). Therefore, on non-Linux systems, we'll go back to
sourcing via the old approach (which I think is inferior in that I don't
think it sets up keyboard completions).