[install-packages-in-background] Include directory in notifications
This morning, I iterated through all of my repos and ran
install-packages-in-background on each. One of the `bundle install` jobs
failed, and I received a notification about it, but I didn't know in
which directory it failed. Adding the directory name to the
notifications should avoid that issue in the future.