[Gemfile] Remove solargraph
Although I'm not sure it's quite as good with respect to jumping to
method definitions etc, for simplicity and for lower resource usage on
my computer, I am now relying on ruby-lsp for go-to-definition
functionality. And, even if I were not to rely on ruby-lsp, I might lean
on https://github.com/pheen/fuzzy_ruby_server as my next
go-to-definition provider, rather than solargraph. So, here we remove
solargraph.