[bin-lint] Introduce bin-lint/ directory (initially with gitleaks) (#30)
The idea of this directory is that, for linting scripts that don't vary
at all between repos, we can just point the repo to run the scripts that
will live (in a DRY fashion) here.

For linting steps that do vary between repos, though, we'll continue to
have those linting scripts live in the individual repos.