[background] Rename from background-and-silence
`background-and-silence` sounds sort of like it can't communicate at
all, but a script backgrounded in this way can still communicate out,
e.g. via `bin/perform-background-step`. I think that the assumption is
sort of that we wouldn't want a process that we background to be
printing random output to the terminal from which it was launched, so
let's just leave that part implicit.