[wait-for-gh-checks] Wait an initial 5 seconds [DOT-46] (#66)
If tests failed in a previous run, and then I push again, there is a
risk that the old checks will still be considered active, and
`wait-for-gh-checks` will fail immediately due to the previously failed
checks. To mitigate against this, wait for an initial 5 seconds before
performing the first check.