[zshrc] Append (not prepend) yarn global bin to $PATH
When prepending, this was putting `/usr/local/bin` in the $PATH ahead of
the rbenv shims directory. Since `/usr/local/bin` contains an old `ruby`
executable, that was overriding the rbenv shims. This change gives rbenv
control of our `ruby` again. Unfortunately, I'm not sure what changed
such that this is now necessary.