[check-git-push-readiness] Rename from check-git-push-safety
I am going to expand the checks herein to check for whether git has been
configured to respect a `.git-blame-ignore-revs` file, if present. That
doesn't really have anything to do with push readiness, but it's
something that I want to check when using git, and this is a reasonably
good place to add that check.