[bolso/prompt] Add git ahead and behind counts [DOT-58]
Also, delete a lot of stuff we aren't actually using.

I might do this somewhat differently if I had full control, but, if I
want to integrate with the oh-my-zsh theming system, then I don't have
full control. The advantage of using the oh-my-zsh theming system is
that I think it allows for asynchronicity, which should improve
performance, which is an important benefit. Overall, I think this is
good enough.