[bolso/prompt] Add git ahead and behind counts [DOT-58] (#115)
Also, delete a lot of stuff we aren't actually using.

I might do this somewhat differently if I had full control, but, if I
want to integrate with the oh-my-zsh theming system, then I don't have
full control. The advantage of using the oh-my-zsh theming system is
that I think it allows for asynchronicity, which should improve
performance, which is an important benefit. Overall, I think this is
good enough.

# Screenshot


![image](https://github.com/user-attachments/assets/7db21fb3-e58a-4bbc-838e-5b3d7af42448)