[Brewfile] Add git
Most systems probably come pre-installed with a version of git, but I
like to install via Homebrew (to be able to easily get the latest
version and to stay up to date). Also, for example, the `asdf` setup
instructions include `brew install coreutils curl git`, so, if I naively
follow those, as I am wont to do, then I'll end up installing git via
Homebrew for that reason, anyway. So, might as well make it official.