[cpy] Add support for a file argument [DOT-61]
Up until now, `cpy` only supported copying piped input from stdin. This
change adds support for another option, providing a file name argument
(in which case the file's contents will be copied to the clipboard).

I'm also adding a `file-arg-content-or-stdin` executable, which does
what it says on the tin.

I'm also adding a `bytes` command that will print the decimal ASCII
character codes of any text piped into it. I used this `bytes` command
to debug some of the issues related to newlines in `cpy` and
`file-arg-content-or-stdin`.