[runger-config] Add option to print current config [DOT-64] (#140)
I'm sort of dismayed at how difficult this was / how little there is in
the Lua standard library. This makes me less inclined to use Lua as a
scripting language, going forward. I guess maybe the reason that it's as
fast as it is is because it's so minimal. Maybe a scripting language
that is both full featured and fast (including startup time) does not
exist.