[runger-config] Rewrite in Crystal (from Lua)
Also, create a framework to manage the automatic compilation of Crystal
programs when executed.

This change also tweaks the zshrc to add to the PATH in an easier to
read and more maintainable way, by adding to a `path` array. I'm not
quite sure why/how this works (since we are referring to a `path` array
rather than the `PATH` string), but it does.

This change also removes `$HOME/bin` as a PATH directory, since I don't
currently use it. Instead, we are adding a subdirectory of that
directory to the path, `$HOME/bin/crystal-symlinks`.