[kitty_opener] Updated editor path from dotfiles-personal to dotfiles (#154)
I moved it recently.

Also, uncomment the `debug` method. I don't think it's a big problem to
have it always uncommented, and it saves the hassle of uncommenting it
whenever I am debugging something.

Also, add a test for the situation where I encountered this issue (since
I thought there was a possibility that might be related to the problem).
The test was passing even without any other changes, but I figure that
we might as well add it, anyway.