[zshrc] Use env vars to avoid symlink loops in Crystal repos [DOT-69]
(... such as this one.)

https://github.com/crystal-lang/shards/issues/ 198
https://stackoverflow.com/a/44674718/4009384
https://man.archlinux.org/man/shards.1.en#ENVIRONMENT_VARIABLES

This is definitely a little bit hacky, but it seems to work. Instead of
installing Crystal shards to `./lib/` in the current directory, we will
instead install them outside of the project, in
`$HOME/.shards/$(basename $PWD)/`.

**Motivation:** having the shards within the project directory causes
problems because there are symlink loops within the installed shards, as
illustrated by this partial `tree` view (shown from the new install
location in `~/.shards/`):

```
~/.shards     11:14:07 PM
$ tree dotfiles
dotfiles
├── clim
│   ├── lib -> ..
│   ├── LICENSE
│   │ [...]
```

As you can see, `clim/lib` is a symlink back up to the parent directory
(`..`), which I guess comes of as a "file system loop". This causes
problems for some tooling:

- breaks creating a new file through my VS Code extension
- causes Tailwind VS Code extension to try to use 100% CPU forever
- causes a warning when searching with e.g. rg -Luu '\b(not-there)\b'

These problems aren't really fundamentally resolved by this change -- I
think we'll still have these problems if using the aforementioned tools
in the `~/.shards/` directory -- but, at least, this way, the problems
are removed from the repos in which I most actively develop. I'm
relatively unlikely to open `~/.shards/` in VS Code or to search it with
`rg`.

Also, add `--warnings=none` to the `crystal build` in
`run-crystal-program`. I started to see a bunch of compilation warnings
when e.g. compiling `runger-config` via `rm -rf ~/bin`, new terminal
tab, and `runger-config -s`. I am guessing that is somehow caused by
having moved the shards out of the project repo. `--warnings=none`
re-suppresses these warnings.