Revert "[Brewfile] Add doppler (and its tap and gnupg dependency)"
This reverts commit 6a5f146a48f3304d87ccee5a7268bd0e45c471f9.

I decided not to use Doppler, because as of
https://github.com/davidrunger/david_runger/pull/4879 we no longer need
access to any secrets in order to build the Docker image for the Rails
app, and nor do we need to add `.env.production.local` to the built
image.

Thinking that the opposite of those things was true had been one of my
main motivations for moving to Doppler, so now there is not much value
in moving to Doppler, I think, as much as I think it's a pretty great
service/tool.