[crystal] Set SHARDS_INSTALL_PATH via wrapper (not zshrc) (#188)
The problem is that zshrc only runs at terminal startup (unless later
manually sourced). So, if I `cd` into a different directory and run
`shards install`, then, using the previous zshrc-based approach, the
SHARDS_INSTALL_PATH would be outdated/wrong. This new approach ensures
that `SHARDS_INSTALL_PATH` is set as desired for the current directory,
even if the current directory is not the same directory as we were in
when the terminal was started.