[install] Remove '~/.rubocop.yml'
Having this file causes excessive CPU usage when opening `typelizer` in
VS Code (maybe only after running `spec/app/bin/setup` in that repo?).

This seems to be because of Ruby LSP. I have set RuboCop as my Ruby
formatter in VS Code, so Ruby LSP tries to load my `~/.rubocop.yml`, but
that seems to error and get hung up in an endless, CPU-burning loop,
having something to do with `runger_style` not being available.

I'm not sure that I'm getting a ton of value from having this
global/home `~/.rubocop.yml`, so, to avoid the above issue (and similar
issues that might occur in other similarly configured repos), let's try
removing it, and see if I miss it.