Move linux/mac aliases to dedicated files
Motivation: This allows us to include the appropriate OS-specific
aliases into the main `aliases.zsh` file. This, in turn, makes the
results of `wh` more accurate. Without this change, `wh fx` says that
`fx` is the Homebrew-managed interactive JSON explorer terminal program,
whereas reality and `which` say that is is an alias for
`firefox-devedition`. `wh` sources the `aliases.zsh` file, though, which
(as of now) in turn sources the OS-specific aliases file, so now `wh`
can accurately report that it is an alias for `firefox-devedition`
rather than inaccurately reporting that it is the JSON explorer program.