[lint/gitleaks] Update from deprecated 'detect' command to 'git'
I think that there is a bug with `detect` in at least the latest version
(8.23.2) where it has stopped respecting the `--log-opts` option?
Switching from `gitleaks detect` to the new, not-deprecated `gitleaks
git` command seems to avoid/fix that issue. And it's probably good, in
general, anyway, to update to that non-deprecated command.