[wait-for-gh-checks] Handle checks with 'ERROR' status
The status of the Percy check was `"ERROR"` after Percy detected a diff.

Also, tweak how we count `total_num_started_checks`. Instead of
whitelisting statuses for inclusion in the count, we will now instead
blacklist statuses that we don't want to include in the count. This
seems to me slightly less likely to require maintenance/updates in the
future. Today's change, with the old approach, would have necessitated
adding `ERROR` to the `ALL_STATUSES` list. With this change, we can
delete the `ALL_STATUSES` list and no longer need to maintain it. The
`ALL_STATUSES` name was misleading, anyway. A more accurate name would
have been `ALL_NON_SKIPPED_STATUSES` or some such.