[merge-prs] Don't sleep after merging last PR in each repo (#423)
The sleeps are to allow dependabot time to check whether any other PRs
need to be rebased and to update those PRs with a message indicating
such. There is no point in waiting for that after merging the last PR,
though, because there are no other PRs that we want to avoid merging
prematurely by not giving dependabot time to do that.