Move --quit-if-one-screen option to front of LESS (#432)
Reason: Ruby's OptionParser will try to append `Fe` to the LESS
environment variable, if it's defined. When `--quit-if-one-screen` is at
the end, this results in the error `There is no quit-if-one-screenFe
option`. By moving the verbose `--quit-if-o`ne-screen` flag to the front
and bringing `-X` to the back, that short form flag can be combined with
the `-F` and `-e` options added by Ruby without issue.