[rebase-branch-with-remote] Remove check for no uncommitted changes (#441)
I'm not sure why I have this?, and I don't think it provides much
benefit. It added a lot of complexity to `gfcob`, which I'm able to
remove with this PR. But even those complexities didn't handle all
situations, such as when a new file is being created (which doesn't get
stashed and so which still ends up blocking the rebase).