[stack-trace-viewer] Print code in reverse of stack trace [DOT-87]
Ruby prints the most recently called line at the top. When
`stack-trace-viewer` also prints the code context of the stack trace in
this same order, then I have to scroll all the way back up (which is
sometimes pretty far) to view the "top" lines of the stack trace, which
are usually the lines of greatest interest. Therefore, instead, this
change prints the more recently called lines last, rather than first.
Thus, the code of the most recently called lines will be visible without
needing to scroll.