Exclude square brackets characters from file paths [DOT-91]
```
![image](https://github.com/user-attachments/assets/c9dd2b93-9bd9-4c59-b6eb-9b3533285c7b)
```

This change fixes a bug wherein, for the above text, our kitty hints
selector would create a hint on the `i` in `image`, and not for the
`github` URL starting with `https://...` .

This is accomplished by modifying our regex that matches file paths (and
other text) to not consider square brackets to possibly be part of
filenames. In reality, on Linux, I think that square brackets _can_ be
part of filenames. But, in practical terms, I think that's unlikely, and
I think that having better behavior on a known case such as mentioned
above is probably worth the cost of theoretically not catching a file
path with square brackets.

Also, apply auto-formatting (via ctrl-shift-r in my editor) to the
`kitty_opener_test.py` file.