Omit PR description from commit message for dependabot PRs [DOT-92]
Sometimes, I check out and iterate on a Dependabot PR, and then I
ultimately merge that PR via `wait-merge`. It would be nice if, in these
situations, `wait-merge` would not merge the PR description as the
commit body. (Note: we already do this in `merge-prs`.) This change
implements that functionality.