[github-stars-explorer] Increase cache TTL from 1 week to 30 days (#509)
This data doesn't really need to be fresh. If I want it to be fresh,
then I can clear the cache manually. By increasing the cache time,
though, this at least gives me the option to re-use (potentially
somewhat stale) data, if I'm okay with that, which I often will be. If
the cache TTL is short, then I don't have a choice; it simply expires.