Rename 'mise.toml' to 'mise.source.toml' (#542)
With the file named `mise.toml`, we get this warning:

```
mise WARN  Config files in ~/code/dotfiles/mise.toml are not trusted.
Trust them with `mise trust`. See https://mise.jdx.dev/cli/trust.html for more information.
```

We don't want this file to be treated as an actual config file. We just
want it to be the source for the symlink to an actual (trusted) config
file. Renaming the file accomplishes this.