[github-stars-explorer] Allow exploring specific repo(s) [DOT-96]
I found the results pretty interesting when doing this for
`laserlemon/rspec-wait` (by just temporarily modifying the source code
to only use that repo). It might be useful to make it easier to explore
the stars of the stargazers of other repos. This change accomplishes
that, by supporting the specification of a `REPOS_TO_EXPLORE` env var.