[github-stars-explorer] Allow exploring specific repo(s) [DOT-96] (#569)
I found the results pretty interesting when doing this for
`laserlemon/rspec-wait` (by just temporarily modifying the source code
to only use that repo). It might be useful to make it easier to explore
the stars of the stargazers of other specific repos. It sort of seems
like I get a better signal from somewhat lesser starred repos (which
`laserlemon/rspec-wait` is an example of). I think that this is because
it probably takes a little more taste and discernment to star a lesser
known repo, so those stargazers have relatively better taste, on
average.

This change implements support for exploring specific repo(s) via a
`REPOS_TO_EXPLORE` env var.