[gal] Remove parenthetical description of 'ptest' from '--help'
It's not really necessary. I am the only person who uses this tool, and
I know what it means. I think that the clutter is not worthwhile. Also,
now we have `ctest`, so we should probably have parenthetical
descriptions for both `ctest` and `ptest` or neither. I think having it
for both is too much / too verbose, so this change goes with neither.