[active-branches] Handle regex characters (e.g. '+') in branch names
... by using `awk` rather than `rg`.

This fixes a bug. When I was on a branch called
`avoid-n+1-queries-in-version_limit-when-destroying`, `gc` (which uses
`active-branches`) was giving me the current branch
(`avoid-n+1-queries-in-version_limit-when-destroying`) as an option to
switch to, because the `+` character in the branch name was being
interpreted as a regex metacharacter, rather than being essentially
interpreted literally. So, that current branch wasn't being excluded
from the `active-branches` list. This change fixes the issue by just
using `awk` in such a way that we don't need to worry about escaping the
current branch name as a regex and instead we can just check for
equality directly.