[update-npm-packages] Use npm-check-updates for pnpm repos (#591)
https://x.com/i/grok?conversation=1901911782148460993

The problem with `pnpm update --latest` is that it will use alpha
versions, which I don't really want.

`npm-check-updates` will bump across major versions (which I want) but
it won't use alpha versions (which I don't want to use), so it's
perfect.

Side note: the `--interactive` flag is pretty nice to use, but not
applicable here, I guess.