[upload-png-to-s3] Add script to upload a PNG to S3 (#612)
I think that I'm going to increasingly host my app's images on S3 rather
than in the git repo. To this end, I want to make it easy to upload
images. For now, I'll start with this PNG-specific uploader. It's
PNG-specific because of the `Content-Type` response header that it sets.