[update-npm-packages] Update transitive dependencies w/ 'pnpm update'
I think that `pnpx npm-check-updates --upgrade && pnpm install` doesn't
update transitive dependencies for which there is no update available
for the package that we reference in our `package.json`. In contrast, I
think that `pnpm update` will update these dependencies. So, let's run
`pnpm update`, in addition to `pnpx npm-check-updates --upgrade && pnpm
install`, to also update transitive dependencies for which there is no
available for update for the package that we have in `package.json`.