[gac] Move from an alias to an executable script
One advantage of this is that it better responds to env var
customizations. For example, when it was an alias, `GIT_EDITOR=nvim gac`
would open my default editor to edit the commit message. I think that
this is because it expanded to `GIT_EDITOR=nvim ga . && gcom` and so the
`GIT_EDITOR` env var only applied to the `ga .` part and not to the
`gcom` part.

In contrast, when it's a script, `GIT_EDITOR=nvim gac` does use `nvim`
to edit the commit message, since the `GIT_EDITOR` env var is applied to
not only the `ga .` part of the command, but also the `gcom` part.